# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "e-book api with drizzle",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api"
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["orders"],
        "summary": "get orders",
        "description": "get orders",
        "operationId": "getOrders",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["orders"],
        "summary": "create order",
        "description": "create order",
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": {
                    "type": "string",
                    "example": "a73bde09-1f83-483f-8c66-a31b79f99ab1"
                  },
                  "paymentMethodId": {
                    "type": "string",
                    "example": "3af6af84-5ade-4563-a0ec-2e333e9dcbf7"
                  },
                  "quantity": {
                    "type": "number",
                    "example": 2
                  },
                  "shippingServiceId": {
                    "type": "string",
                    "example": "1767a604-8905-484b-9691-cff17c2991b2"
                  }
                }
              },
              "examples": {
                "create order": {
                  "value": {
                    "bookId": "a73bde09-1f83-483f-8c66-a31b79f99ab1",
                    "paymentMethodId": "3af6af84-5ade-4563-a0ec-2e333e9dcbf7",
                    "quantity": 2,
                    "shippingServiceId": "1767a604-8905-484b-9691-cff17c2991b2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}